Use the '-mmee-machine' GCC argument
This matches how users will actually use the compiled MEE libraries.  As
a result we have to generate a linker script (which requires another
program) as well as link against the C library (to pull in the
initialization code).

This resulted in finding a handful of small bugs in naming conventions
and makefile dependencies.

I've successfull built this with

    autoreconf -i
    ./configure \
        --prefix=$(pwd)/install \
        --host=riscv64-sifive-mee \
        --with-machine-name=qemu-sifive_e-32 \
        --with-machine-dts=$(pwd)/dts/qemu-sifive_e-32.dts \
        --with-mee-header-generator=TOOLSfreedom-mee_header-generator \
        --with-ldscript-generator=TOOLSfreedom-ldscript-generator
    make
        CCASFLAGS="-Og -march=rv32imac -mabi=ilp32 -g" \
        CFLAGS='-Og -march=rv32imac -mabi=ilp32 -mcmodel=medany -mexplicit-relocs -g' \
        LDFLAGS="-march=rv32imac -mabi=ilp32 -g" \
        check

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>