Respect the HFROSC's divider
Note that this still doesn't handle the trim, which is a bit harder.  As
a result the clock is extremely unstable.  We should really have the
HiFive1 default to the HFXOSC, ideally via the device tree's "/chosen"
node.

For now this at least gets you to within about 10%, which is
something...

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>