Move from riscv64-sifive-mee to riscv64-sifive-elf
I've added support to our ELF toolchain that replaces the various
features of the MEE toolchain we used to have in a cleaner way.  This
allows us to both ship a single toolchain binary as well as remove some
command line arguments that were required by the MEE toolchain.

This simplifies the MEE's build scripts, and should also simplify users
build scripts as well.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>