Add export of standalone project
Split out board init, toolchain setup, and program build into
scripts/standalone.mk which can be included in the top-level Makefile to
allow in-SDK builds or be exported and used on its own.

Create a `standalone` Make target which copies and example, the built
Freedom Metal library, and the standalone Makefile into an external
project.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>