Build metal libs into install/lib/CONFIGURATION
Removes the INCLUDE_METAL_SOURCES option. If we want to be able to
support multiple build configurations for standalone projects, the Metal
library must be rebuilt to take into account those options.

An alternative is to put back the prebuilt library option and just
remove the configuration option from standalone projects without the
Metal library sources.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>