Fix PMP address generation
The position of the least-significant zero sets the range, so there was
an off-by-one error.

Given a protected range of 128 bytes, we want the zero to be at bit

    log2(128 >> 2) = 5

which means we need to and the address with 0b...1101111, which we
calculate with

    (123 >> 3) = 0b10000

So we actually need to shift the requested NAPOT_SIZE by 3, not 2.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>