Create settings.mk if it doesn't exist at first run
This is the last change needed to completely reproduce BSPs from only a
design.dts. With Freedom Devicetree Tools and Devicetree Compiler in the
path and only bsp/foo/design.dts, we can generate the entire bsp, build
metal, and build and link the program with only `make TARGET=foo
software`.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>