doc: fix an error if the doc is not generated from github
The commit 57c45c04 expect Sphinx to find the environment variable
RELEASE_TAG.
It seems it's only the case when the documentation is generated from
github using the workflow defined in the release.yml file.

This commit allows to generate the doc from outside github, by setting
version to the current commit id if we are on a git repository, and
"n/a" otherwise.