unmatched: change timebase-frequency to 1.2GHz
Changes the timbebase-frequency in DT file and updates the BSP files.
The timebase-frequency should normally represent RTC clock, but
timebase-frequency is used as clock frequency in FESDK now, so we
change it to 1.2GHz here for the time being. This is just a workaround
patch, we have had a plan to fix the issue in FESDK, so it would be
corrected again later.

Signed-off-by: Zong Li <zong.li@sifive.com>